Move common code to shared
